+++ ACKategories-iOSTests/EdgeInsetsTests.swift
@@ -14,4 +14,16 @@ final class EdgeInsetsTests: XCTestCase {
             )
         )
     }
+    
+    func test_zero() {
+        XCTAssertEqual(
+            EdgeInsets.zero,
+            .init(
+                top: 0,
+                leading: 0,
+                bottom: 0,
+                trailing: 0
+            )
+        )
+    }
 }